requirements.txt
Django>=3.2,<3.3
wagtail>=2.13,<2.14
wagtail-generic-chooser
